+++ include/profiles.h
@@ -1,8 +1,6 @@
 #ifndef PROFILES_H
 #define PROFILES_H
 
-#include <stdlib.h>
-
 #include <gsl/gsl_interp.h>
 #include <gsl/gsl_dht.h>
 
@@ -51,8 +49,8 @@ inv_profile_e;
 
 typedef struct
 {//{{{
-    size_t start; // the start index in the signal grid
-    size_t len;   // length of this batch
+    long start; // the start index in the signal grid
+    long len;   // length of this batch
     int incr;     // +-1 --> loop over signal grid such that
                   //    theta is always decreasing
     double *data; // either t_of_s or dtsq_of_s, of length len
